Update interaction and improve code readability
After every command to stop the head, a command to make a normal
face expression is issued so that situations where the robot is
stopped midblink are taken care of. The intensities and timings
of the expressions where revised as well as the timings between
expressions. The parts of the interaction that repeat often where
encapsulated in functions to ease readability.